KSCU-7: Begin to modularize KlaviyoRecreate-Cart controller and associated logic on SiteGen side. Create new helper file in klaviyo_core to make logic accessible between SiteGen & SFRA as needed.
